+++ packages/react/src/sds/surveys/SurveyAnsweringForm/types.ts
@@ -27,11 +27,8 @@ export type SurveyFormSubmitResult =
 
 export type SurveyAnsweringFormMode = "stepped" | "all-questions"
 
-export interface SurveyAnsweringFormProps {
+interface SurveyAnsweringFormBaseProps {
   elements: SurveyFormBuilderElement[]
-  onSubmit: (
-    answers: SurveySubmitAnswers
-  ) => Promise<SurveyFormSubmitResult> | SurveyFormSubmitResult
   mode: SurveyAnsweringFormMode
   title: string
   isOpen: boolean
@@ -40,7 +37,31 @@ export interface SurveyAnsweringFormProps {
   allowToChangeFullscreen?: boolean
   defaultValues?: Partial<SurveyAnswers>
   errorTriggerMode?: F0FormErrorTriggerMode
+  loading?: boolean
+  labels?: {
+    empty?: {
+      title?: string
+      description?: string
+      emoji?: string
+    }
+  }
 }
+
+export interface SurveyAnsweringFormDefaultProps extends SurveyAnsweringFormBaseProps {
+  preview?: false
+  onSubmit: (
+    answers: SurveySubmitAnswers
+  ) => Promise<SurveyFormSubmitResult> | SurveyFormSubmitResult
+}
+
+export interface SurveyAnsweringFormPreviewProps extends SurveyAnsweringFormBaseProps {
+  preview: true
+  onSubmit?: never
+}
+
+export type SurveyAnsweringFormProps =
+  | SurveyAnsweringFormDefaultProps
+  | SurveyAnsweringFormPreviewProps
 
 export type FlatQuestion = {
   id: string